% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal-anchor-nn.R
\name{apply_gradient_a}
\alias{apply_gradient_a}
\title{anchor ontram gradient}
\usage{
apply_gradient_a(
  x_train,
  y_train,
  model,
  img_train = NULL,
  verbose = FALSE,
  response_varying = FALSE,
  xi = 0,
  prm = 0
)
}
\description{
anchor ontram gradient
}
