% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colr-anchor-exact-right.R
\name{colr_gradient_r}
\alias{colr_gradient_r}
\title{anchor colr gradient}
\usage{
colr_gradient_r(
  x_train,
  y_train,
  yp_train,
  cens_train,
  model,
  img_train = NULL,
  verbose = FALSE,
  response_varying = FALSE,
  xi = 0,
  prm = 0
)
}
\description{
anchor colr gradient
}
