% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colr-anchor-exact-right.R
\name{colr_loss_r}
\alias{colr_loss_r}
\title{Penalized log-likelihood for the c-logit model}
\usage{
colr_loss_r(
  gammas,
  beta = NULL,
  eta = NULL,
  y,
  yp,
  cens,
  prm,
  distr = k_sigmoid,
  xi,
  indiv = FALSE
)
}
\description{
Penalized log-likelihood for the c-logit model
}
\examples{
library(tram)
library(survival)
data("BostonHousing2", package = "mlbench")
BostonHousing2$cmedvv <- with(BostonHousing2, Surv(cmedv, cmedv != 50))
m <- Colr(cmedv ~ 1, data = BostonHousing2, order = 4)
mn <- Colr(cmedvv ~ nox, data = BostonHousing2, order = 4)
trdat <- tramnet:::.get_tram_data(m)
ntheta <- trdat$npar
y <- trdat$exact$ay
yp <- trdat$exact$aypr
x <- model.matrix(~ 0 + nox, data = BostonHousing2)
mbl <- mod_baseline(ntheta + 1L)
msh <- mod_shift(1L)
theta_to_gamma <- function(gammas) {
  thetas <- gammas
  thetas[2:length(thetas)] <- log(diff(gammas))
  thetas[is.nan(thetas)] <- - 1e20
  return(thetas)
}
gms <- theta_to_gamma(coef(mn, with_baseline = TRUE)[-(ntheta + 1)])
tmp <- get_weights(mbl)
tmp[[1]][] <- 0
tmp[[1]][] <- gms
set_weights(mbl, tmp)
tmp2 <- get_weights(msh)
tmp2[[1]][] <- coef(mn)
set_weights(msh, tmp2)
mo <- ontram(mbl, msh, y_dim = ntheta, x_dim = 1L, n_batches = 1, epochs = 1e3, lr = 0)
gammas <- mbl(matrix(1.0, ncol = 1L, nrow = 1L))
betas <- k_squeeze(msh(x), axis = 0L)
cens <- as.numeric(BostonHousing2$cmedv == 50)
colr_loss_r(gammas, betas, NULL, k_constant(y), k_constant(yp), k_constant(cens), prm = 0, xi = 0)
- logLik(mn) / nrow(BostonHousing2)
coef(mn, w = T); coef(mo, w = T)
colr_gradient_r(k_constant(x), k_constant(y), k_constant(yp), k_constant(cens), mo, verbose = TRUE)
logLik.colrnn_r(mo, x = x, y = y, yp = yp, cens = cens)
}
