% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal-anchor-nn.R
\name{compute_logLik_a}
\alias{compute_logLik_a}
\title{anchor logLik ontram}
\usage{
compute_logLik_a(
  gammas,
  beta = NULL,
  eta = NULL,
  y_train,
  prm,
  distr = k_sigmoid,
  xi
)
}
\description{
anchor logLik ontram
}
\examples{
y <- k_constant(matrix(c(0, 0, 1, 0, 0, 0,
                         0, 1, 0, 0, 0, 0), nrow = 2L, ncol = 6L, byrow = TRUE))
gammas <- k_constant(matrix(c(-2, 0, 0, 0, 0), nrow = 1L, ncol = 5L, byrow = TRUE))
prm <- k_constant(matrix(c(0.46, 0.04, 0.04, 0.1), nrow = 2L, ncol = 2L, byrow = TRUE))
compute_logLik_a(gammas, y_train = y, prm = prm, xi = 10)
}
