% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colr-anchor-exact-right.R
\name{fit_colr_r}
\alias{fit_colr_r}
\title{anchor fit colr}
\usage{
fit_colr_r(
  model,
  history = FALSE,
  x_train = NULL,
  y_train,
  yp_train,
  cens_train,
  img_train = NULL,
  save_model = FALSE,
  x_test = NULL,
  y_test = NULL,
  img_test = NULL,
  xi,
  a_train,
  verbose = 1
)
}
\description{
anchor fit colr
}
\examples{
library(tram)
library(survival)
data("BostonHousing2", package = "mlbench")
BostonHousing2$cmedvv <- with(BostonHousing2, Surv(cmedv, cmedv != 50))
m <- Colr(cmedv ~ 1, data = BostonHousing2, order = 4, log_first = TRUE)
mn <- Colr(cmedvv ~ rm, data = BostonHousing2, order = 4, log_first = TRUE)
trdat <- tramnet:::.get_tram_data(m)
ntheta <- trdat$npar
y <- trdat$exact$ay
yp <- trdat$exact$aypr
x <- model.matrix(~ 0 + rm, data = BostonHousing2)
mbl <- mod_baseline(ntheta + 1L)
msh <- mod_shift(1L)
theta_to_gamma <- function(gammas) {
  thetas <- gammas
  thetas[2:length(thetas)] <- log(diff(gammas))
  thetas[is.nan(thetas)] <- - 1e20
  return(thetas)
}
gms <- theta_to_gamma(coef(mn, with_baseline = TRUE)[-(ntheta+1)])
tmp <- get_weights(mbl)
tmp[[1]][] <- 0
tmp[[1]][] <- gms
set_weights(mbl, tmp)
tmp2 <- get_weights(msh)
tmp2[[1]][] <- coef(mn)
set_weights(msh, tmp2)
mo <- ontram(mbl, msh, y_dim = ntheta, x_dim = 1L, n_batches = 1, epochs = 1e3, lr = 1e-3)
A <- model.matrix(~ town, data = BostonHousing2)
cens <- as.numeric(BostonHousing2$cmedv == 50)
fit_colr_r(mo, x_train = x, y_train = y, yp_train = yp, cens_train = cens, a_train = A, xi = 0)
coef(mn, w = T); coef(mo, w = T)
}
