% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal-anchor-nn.R
\name{fit_ontram_a}
\alias{fit_ontram_a}
\title{anchor fit ontram}
\usage{
fit_ontram_a(
  model,
  history = FALSE,
  x_train = NULL,
  y_train,
  img_train = NULL,
  save_model = FALSE,
  x_test = NULL,
  y_test = NULL,
  img_test = NULL,
  xi,
  a_train,
  verbose = 1
)
}
\description{
anchor fit ontram
}
