% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{gen_dat_ordinal}
\alias{gen_dat_ordinal}
\title{ordinal IV scenario}
\usage{
gen_dat_ordinal(
  n = 1000,
  sd = 1,
  ba = 1,
  bx = 1,
  bh = 1,
  perturb = FALSE,
  shift = 1.8,
  nclasses = 4
)
}
\description{
ordinal IV scenario
}
