% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colr-anchor-exact-right.R
\name{logLik.colrnn_r}
\alias{logLik.colrnn_r}
\title{Function for predicting cdf, pdf and class and compute logLik}
\usage{
\method{logLik}{colrnn_r}(model, x = NULL, y, yp, cens, im = NULL, indiv = FALSE)
}
\description{
Function for predicting cdf, pdf and class and compute logLik
}
