% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colr-anchor-exact-right.R
\name{logLik.colrnn_r}
\alias{logLik.colrnn_r}
\title{Out-of-sample logLik for the c-logit model}
\usage{
\method{logLik}{colrnn_r}(model, x = NULL, y, yp, cens, im = NULL, indiv = FALSE)
}
\description{
Out-of-sample logLik for the c-logit model
}
